4th mod

+++ include/rand_num.h
@@ -1,4 +1,4 @@
 #ifndef RAND_NUM_H_INCLUDED
 #define RAND_NUM_H_INCLUDED
-	float rand_num(float);
+	double rand_num(double);
 #endif
