+++ include/viterbi_init.h
@@ -1,6 +1,6 @@
 #ifndef VITERBIINIT_H_INCLUDED
 #define VITERBIINIT_H_INCLUDED
 
-void viterbi_init(float *FI, float *A_Vect, float *LogPdf, int N, int T);
+void viterbi_init(double *FI, double *A_Vect, double *LogPdf, int N, int T);
 
 #endif
