+++ include/wdensity.h
@@ -1,6 +1,6 @@
 #ifndef WDENSITY_H_INCLUDED
 #define WDENSITY_H_INCLUDED
 
-void wdensity(float *Mu_Vect, float *Sigma_Vect, float *PComp_Vect, double *Test_Vect, float* Log, int D, int N, int K, int T);
+void wdensity(double *Mu_Vect, double *Sigma_Vect, double *PComp_Vect, double *Test_Vect, double* Log, int D, int N, int K, int T);
 
 #endif
